requirements.txt
streamlit
tensorflow
keras
numpy
pillow
opencv-python-headless
matplotlib
scikit-learn